requirements.txt
requests>=2.28.0
flask>=2.0.0
pandas>=1.4.0
faker>=18.0.0
python-dotenv>=0.19.0
schedule>=1.1.0
cryptography>=3.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
openpyxl>=3.0.0
jinja2>=3.0.0
